+++ container-engine-lib/lib/backend_interface/objects/service/service_registration.go
@@ -5,17 +5,20 @@ import (
 	"net"
 )
 
-// A ServiceRegistration is a stub for a soon-to-be-started service
-// We have this as an independent object so we can return the container's IP
-// address to the user before the container exists
+// A ServiceRegistration is a stub for a started service
+// We had created this to return the user's IP before the container exists
+// The user doesn't need the private IP address anymore.
+// Now this is used for caches in APIC & DockerKurtosisBackend
+// Also partitioning in APIC is based on the private IP address this returns
+// TODO visit removing this  after partitioning is moved to container-engine-lib and data is stored in a database
 type ServiceRegistration struct {
-	id				 ServiceID
-	guid             ServiceGUID
-	enclaveId        enclave.EnclaveID
+	id        ServiceID
+	guid      ServiceGUID
+	enclaveId enclave.EnclaveID
 
 	// The private IP is the IP of the service within the enclave, meaning other services can use this IP to communicate
 	// with the service
-	privateIp        net.IP
+	privateIp net.IP
 }
 
 func NewServiceRegistration(id ServiceID, guid ServiceGUID, enclaveId enclave.EnclaveID, privateIp net.IP) *ServiceRegistration {